+++ src/index.js
@@ -120,7 +120,7 @@ Async.auto({
     if (err) {
         // Log a fatal error and exit.
         // We need this to initialize successfully without any errors.
-        Log.fatal(err);
+        Log.fatal({ err: err, additional: err }, 'A fatal error caused the daemon to stop abruptly.');
         process.exit(1);
     }
     rfr('src/http/routes.js');